NEW: redStraightIntake - redStraight but with the intake multiple glyph test instead. No color sensors.
CHANGES:
IntakeRemote - Changed the name it'll show up with to just Remote.
Remote - Disabled within the app. Will no longer show up.
CheckPicto - Added the servo that holds the blocks into the slide.
CheckPictoIntake - Added the servo that holds the block into the slide.
redStraight - Added the servo that holds the block into the slide.